move non-entry point headers into detail subdir
