+++ rosidl_generator_c/resource/idl__functions.h.em
@@ -1,4 +1,4 @@
-// generated from rosidl_generator_c/resource/idl__struct.h.em
+// generated from rosidl_generator_c/resource/idl__functions.h.em
 // with input from @(package_name):@(interface_path)
 // generated code does not contain a copyright notice
 @
@@ -12,8 +12,8 @@
 @#######################################################################
 @{
 from rosidl_cmake import convert_camel_case_to_lower_case_underscore
-include_parts = [package_name] + list(interface_path.parents[0].parts) + \
-    [convert_camel_case_to_lower_case_underscore(interface_path.stem)]
+include_parts = [package_name] + list(interface_path.parents[0].parts) + [
+    'detail', convert_camel_case_to_lower_case_underscore(interface_path.stem)]
 header_guard_variable = '__'.join([x.upper() for x in include_parts]) + \
     '__FUNCTIONS_H_'
 include_base = '/'.join(include_parts)
