+++ rosidl_generator_cpp/rosidl_generator_cpp/__init__.py
@@ -31,9 +31,9 @@ from rosidl_parser.definition import UnboundedSequence
 def generate_cpp(generator_arguments_file):
     mapping = {
         'idl.hpp.em': '%s.hpp',
-        'idl__builder.hpp.em': '%s__builder.hpp',
-        'idl__struct.hpp.em': '%s__struct.hpp',
-        'idl__traits.hpp.em': '%s__traits.hpp',
+        'idl__builder.hpp.em': 'detail/%s__builder.hpp',
+        'idl__struct.hpp.em': 'detail/%s__struct.hpp',
+        'idl__traits.hpp.em': 'detail/%s__traits.hpp',
     }
     generate_files(
         generator_arguments_file, mapping,
