+++ rosidl_typesupport_introspection_c/resource/msg__type_support.c.em
@@ -15,6 +15,9 @@ from rosidl_parser.definition import NamespacedType
 include_parts = [package_name] + list(interface_path.parents[0].parts) + \
     [convert_camel_case_to_lower_case_underscore(interface_path.stem)]
 include_base = '/'.join(include_parts)
+include_parts_detail = [package_name] + list(interface_path.parents[0].parts) + [
+    'detail', convert_camel_case_to_lower_case_underscore(interface_path.stem)]
+include_base_detail = '/'.join(include_parts_detail)
 
 header_files = [
     'stddef.h',  # providing offsetof()
@@ -23,8 +26,8 @@ header_files = [
     'rosidl_typesupport_introspection_c/field_types.h',
     'rosidl_typesupport_introspection_c/identifier.h',
     'rosidl_typesupport_introspection_c/message_introspection.h',
-    include_base + '__functions.h',
-    include_base + '__struct.h',
+    include_base_detail + '__functions.h',
+    include_base_detail + '__struct.h',
 ]
 
 function_prefix = message.structure.namespaced_type.name + '__rosidl_typesupport_introspection_c'
